optimization coding and remove wiggle caution message
