--- data/pwd.json
@@ -1,3 +1,0 @@
-{
-  "password": "root456789"
-}
