+++ public/css/base.css
@@ -31,7 +31,7 @@
   background: rgba(0, 0, 0, 0.75);
   visibility: visible;
 }
-.wrapper .sign_page .s_i_page_box, .wrapper .sign_page .s_u_page_box {
+.wrapper .sign_page .page_box {
   text-align: center;
   width: 340px;
   background: #fff;
@@ -40,7 +40,7 @@
   transform: translateX(-50%);
   z-index: 110;
   border-radius: 6px;
-  padding: 15px;
+  padding: 15px 15px 10px 15px;
   display: grid;
   justify-content: center;
   align-items: center;
@@ -50,7 +50,23 @@
   opacity: 0;
   visibility: hidden;
 }
-.wrapper .sign_page .s_i_page_box .top_decorate, .wrapper .sign_page .s_u_page_box .top_decorate {
+.wrapper .sign_page .page_box.si {
+  top: 30px;
+}
+.wrapper .sign_page .page_box.si.page_box_show {
+  visibility: visible;
+  top: 69px;
+  opacity: 1;
+}
+.wrapper .sign_page .page_box.su {
+  top: 108px;
+}
+.wrapper .sign_page .page_box.su.page_box_show {
+  visibility: visible;
+  top: 69px;
+  opacity: 1;
+}
+.wrapper .sign_page .page_box .top_decorate {
   position: absolute;
   left: 0;
   top: 0;
@@ -59,14 +75,14 @@
   height: 10px;
   background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
 }
-.wrapper .sign_page .s_i_page_box h2, .wrapper .sign_page .s_u_page_box h2 {
+.wrapper .sign_page .page_box h2 {
   font-style: normal;
   font-weight: 700;
   font-size: 24px;
   color: #666;
   margin: 10px auto;
 }
-.wrapper .sign_page .s_i_page_box .s_i_exit_icon, .wrapper .sign_page .s_i_page_box .s_u_exit_icon, .wrapper .sign_page .s_u_page_box .s_i_exit_icon, .wrapper .sign_page .s_u_page_box .s_u_exit_icon {
+.wrapper .sign_page .page_box .exit_icon, .wrapper .sign_page .page_box .exit_icon {
   width: 16px;
   height: 16px;
   background-image: url("../images/icon_close.png");
@@ -77,10 +93,10 @@
   transition: 0.4s;
   cursor: pointer;
 }
-.wrapper .sign_page .s_i_page_box .s_i_exit_icon:hover, .wrapper .sign_page .s_i_page_box .s_u_exit_icon:hover, .wrapper .sign_page .s_u_page_box .s_i_exit_icon:hover, .wrapper .sign_page .s_u_page_box .s_u_exit_icon:hover {
+.wrapper .sign_page .page_box .exit_icon:hover, .wrapper .sign_page .page_box .exit_icon:hover {
   opacity: 1;
 }
-.wrapper .sign_page .s_i_page_box input, .wrapper .sign_page .s_u_page_box input {
+.wrapper .sign_page .page_box input {
   font-size: 16px;
   color: #757575;
   padding: 15px;
@@ -88,7 +104,7 @@
   border-radius: 5px;
   margin-bottom: 15px;
 }
-.wrapper .sign_page .s_i_page_box button, .wrapper .sign_page .s_u_page_box button {
+.wrapper .sign_page .page_box button {
   height: 47px;
   background: #448899;
   border-radius: 5px;
@@ -101,10 +117,24 @@
   cursor: pointer;
   transition: 0.4s;
 }
-.wrapper .sign_page .s_i_page_box button:hover, .wrapper .sign_page .s_u_page_box button:hover {
+.wrapper .sign_page .page_box button:hover {
   opacity: 1;
 }
-.wrapper .sign_page .s_i_page_box p, .wrapper .sign_page .s_u_page_box p {
+.wrapper .sign_page .page_box .caution_box {
+  transition: 0.4s;
+  height: 10px;
+  position: relative;
+}
+.wrapper .sign_page .page_box .caution_box .msg {
+  position: absolute;
+  left: 50%;
+  top: 9px;
+  transform: translateX(-50%);
+  color: rgb(208, 35, 35);
+  transition: 0.4s;
+  opacity: 0;
+}
+.wrapper .sign_page .page_box p {
   font-style: normal;
   font-weight: 500;
   font-size: 16px;
@@ -113,34 +143,8 @@
   top: 50%;
   transform: translateY(-50%);
 }
-.wrapper .sign_page .s_i_page_box p.s_i_msg, .wrapper .sign_page .s_i_page_box p.s_u_msg, .wrapper .sign_page .s_u_page_box p.s_i_msg, .wrapper .sign_page .s_u_page_box p.s_u_msg {
-  color: rgb(208, 35, 35);
-  margin-top: 5px;
-  height: 0px;
-  transition: 0.4s;
-  opacity: 0;
-}
-.wrapper .sign_page .s_i_page_box p.s_i_p, .wrapper .sign_page .s_i_page_box p.s_u_p, .wrapper .sign_page .s_u_page_box p.s_i_p, .wrapper .sign_page .s_u_page_box p.s_u_p {
-  margin-top: 6px;
-}
-.wrapper .sign_page .s_i_page_box p .s_i_to_sign_up, .wrapper .sign_page .s_i_page_box p .s_i_to_sign_in, .wrapper .sign_page .s_u_page_box p .s_i_to_sign_up, .wrapper .sign_page .s_u_page_box p .s_i_to_sign_in {
+.wrapper .sign_page .page_box p .to_sign_up, .wrapper .sign_page .page_box p .to_sign_in {
   cursor: pointer;
-}
-.wrapper .sign_page .s_i_page_box {
-  top: 30px;
-}
-.wrapper .sign_page .s_i_page_box.page_box_show {
-  visibility: visible;
-  top: 69px;
-  opacity: 1;
-}
-.wrapper .sign_page .s_u_page_box {
-  top: 108px;
-}
-.wrapper .sign_page .s_u_page_box.page_box_show {
-  visibility: visible;
-  top: 69px;
-  opacity: 1;
 }
 .wrapper nav {
   width: 100%;
