+++ templates/base.html
@@ -12,26 +12,30 @@
 <body>
   <div class="wrapper">
     <div class="sign_page">
-      <div class="s_i_page_box">
+      <div class="page_box si">
         <div class="top_decorate"></div>
-        <h2 class="s_i_h2">登入會員帳號</h2>
-        <div class="s_i_exit_icon"></div>
-        <input class="s_i_email" type="email" placeholder="輸入電子信箱">
-        <input class="s_i_pwd" type="password" placeholder="輸入密碼 需含小寫英文及數字">
-        <button class="s_i_btn" onclick="signIn();">登入帳戶</button>
-        <p class="s_i_msg"></p>
-        <p class="s_i_p">還沒有帳戶？<span class="s_i_to_sign_up" onclick="showSignUp();">點此註冊</span></p>
+        <h2 class="h2 si">登入會員帳號</h2>
+        <div class="exit_icon si"></div>
+        <input class="email si" type="email" placeholder="輸入電子信箱">
+        <input class="pwd si" type="password" placeholder="輸入密碼 需含小寫英文及數字">
+        <button class="btn si" onclick="signIn();">登入帳戶</button>
+        <div class="caution_box si">
+          <p class="msg si"></p>
+        </div>
+        <p class="p si">還沒有帳戶？<span class="to_sign_up" onclick="showSignUp();">點此註冊</span></p>
       </div>
-      <div class="s_u_page_box">
+      <div class="page_box su">
         <div class="top_decorate"></div>
-        <h2 class="s_u_h2">註冊會員帳號</h2>
-        <div class="s_u_exit_icon"></div>
-        <input class="s_u_name" type="text" placeholder="輸入姓名">
-        <input class="s_u_email" type="email" placeholder="輸入電子信箱">
-        <input class="s_u_pwd" type="password" placeholder="輸入密碼 需含小寫英文及數字">
-        <button class="s_u_btn" onclick="signUp();">註冊新帳戶</button>
-        <p class="s_u_msg"></p>
-        <p class="s_u_p">已經有帳戶了？<span class="s_i_to_sign_in" onclick="toggleSignIn();">點此登入</span></p>
+        <h2 class="h2 su">註冊會員帳號</h2>
+        <div class="exit_icon su"></div>
+        <input class="name su" type="text" placeholder="輸入姓名">
+        <input class="email su" type="email" placeholder="輸入電子信箱">
+        <input class="pwd su" type="password" placeholder="輸入密碼 需含小寫英文及數字">
+        <button class="btn su" onclick="signUp();">註冊新帳戶</button>
+        <div class="caution_box su">
+          <p class="msg su"></p>
+        </div>
+        <p class="p su">已經有帳戶了？<span class="to_sign_in" onclick="toggleSignIn();">點此登入</span></p>
       </div>
     </div>
